+++ cli/cli/commands/enclave/inspect/interactive_repls.go
@@ -3,6 +3,7 @@ package inspect
 import (
 	"context"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/table_printer"
 	"github.com/kurtosis-tech/kurtosis-core/commons/enclave_object_labels"
 	"github.com/palantir/stacktrace"
 )
@@ -19,8 +20,8 @@ func printInteractiveRepls(ctx context.Context, dockerManager *docker_manager.Do
 		return stacktrace.Propagate(err, "An error occurred getting interactive REPL containers by labels: '%+v'", interactiveReplLabels)
 	}
 
-	tabWriter := getTabWriterForPrinting()
-	writeElemsToTabWriter(tabWriter, interactiveReplGUIDColHeader)
+	tablePrinter := table_printer.NewTablePrinter(interactiveReplGUIDColHeader)
+
 	sortedContainers, err := getContainersSortedByGUID(containers)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred sorting interactive REPL containers by GUID")
@@ -30,9 +31,11 @@ func printInteractiveRepls(ctx context.Context, dockerManager *docker_manager.Do
 		if !found {
 			return stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", enclave_object_labels.GUIDLabel, container.GetId(), container.GetLabels())
 		}
-		writeElemsToTabWriter(tabWriter, containerGuid)
+		if err := tablePrinter.AddRow(containerGuid); err != nil {
+			return stacktrace.Propagate(err, "An error occurred writing interactive REPL row for container '%v' to the table printer", containerGuid)
+		}
 	}
-	tabWriter.Flush()
+	tablePrinter.Print()
 
 	return nil
 }