+++ cli/cli/commands/enclave/ls/ls.go
@@ -11,6 +11,8 @@ import (
 	"github.com/docker/docker/client"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager/types"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/enclave_manager/enclave_states"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/enclave_state_from_container_state_retriever"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/logrus_log_levels"
 	"github.com/kurtosis-tech/kurtosis-core/commons/enclave_object_labels"
 	"github.com/palantir/stacktrace"
@@ -67,19 +69,52 @@ func run(cmd *cobra.Command, args []string) error {
 		dockerClient,
 	)
 
-	labels := getLabelsForListEnclaves()
-
-	containers, err := dockerManager.GetContainersByLabels(ctx, labels, true)
+	searchLabels := map[string]string{
+		enclave_object_labels.AppIDLabel: enclave_object_labels.AppIDValue,
+	}
+	kurtosisContainers, err := dockerManager.GetContainersByLabels(ctx, searchLabels, true)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting containers by labels: '%+v'", labels)
+		return stacktrace.Propagate(err, "An error occurred getting Kurtosis containers by labels: '%+v'", searchLabels)
 	}
 
-	fmt.Println(enclaveIdColumnHeader)
-	if containers != nil {
-		enclaveIds := getContainersEnclaveIds(containers)
-		for _, enclaveId := range enclaveIds {
-			fmt.Println(enclaveId)
+	kurtosisContainersByEnclaveId := map[string][]*types.Container{}
+	for _, container := range kurtosisContainers {
+		containerLabels := container.GetLabels()
+		enclaveId, found := containerLabels[enclave_object_labels.EnclaveIDContainerLabel]
+		if !found {
+			return stacktrace.NewError(
+				"No '%v' label found for container '%v'; this is a bug in Kurtosis!",
+				enclave_object_labels.EnclaveIDContainerLabel,
+				container.GetId(),
+			)
 		}
+
+		enclaveContainers, found := kurtosisContainersByEnclaveId[enclaveId]
+		if !found {
+			enclaveContainers = []*types.Container{}
+		}
+		enclaveContainers = append(enclaveContainers, container)
+
+		kurtosisContainersByEnclaveId[enclaveId] = enclaveContainers
+	}
+
+	orderedEnclaveIds := make([]string, len(kurtosisContainersByEnclaveId))
+	enclaveStates := map[string]enclave_states.EnclaveState{}
+	for enclaveId, enclaveContainers := range kurtosisContainersByEnclaveId {
+		orderedEnclaveIds = append(orderedEnclaveIds, enclaveId)
+		enclaveState, err := enclave_state_from_container_state_retriever.GetEnclaveState(enclaveContainers)
+		if err != nil {
+			return stacktrace.NewError("An error occurred getting the state for enclave '%v' from the status of its containers")
+		}
+		enclaveStates[enclaveId] = enclaveState
+	}
+	sort.Strings(orderedEnclaveIds)
+
+
+
+	fmt.Fprintln(logrus.StandardLogger().Out, enclaveIdColumnHeader)
+	for _, enclaveId := range orderedEnclaveIds {
+
 	}
 
 	return nil