+++ cli/cli/enclave_manager/enclave_states/enclave_states.go
@@ -1,0 +1,11 @@
+package enclave_states
+
+type EnclaveState string
+
+const (
+	// One or more containers in the enclave is still running
+	Running EnclaveState = "RUNNING"
+
+	// All containers in the enclave are stopped
+	Stopped EnclaveState = "STOPPED"
+)